chore: move Descriptor type alias to oras/types.py
The Descriptor type is now defined alongside container_type in
oras/types.py, the project's established home for cross-cutting
type aliases. oras/copy/descriptor.py re-exports it for backward
compatibility, so existing imports continue to work.
